backend/requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
python-dotenv==1.1.1
pydantic==2.11.9
python-multipart==0.0.20
requests==2.32.5
httpx==0.28.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# HTML 파싱
beautifulsoup4==4.13.5
readability-lxml==0.8.4.1
lxml==6.0.2


google-genai==1.38.0
google-generativeai==0.8.5

